\name{caGeneSetAnalysis-package}
\Rdversion{1.1}
\alias{caGeneSetAnalysis-package}

\docType{package}

\title{
RWebServices interface to the GSEABase and Category packages.
}

\description{
Provides an RWebServices interface to a part of the functionality provided
by the GSEABase and Category packages.
}

\details{
\tabular{ll}{
Package: \tab caGeneSetAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.1.7\cr
Date: \tab 2009-05-01\cr
License: \tab caBIG\cr
LazyLoad: \tab yes\cr
}
}

\author{
Tom Milac <tmilac@fhcrc.org>
}

%\references{
%Literature or other references for background information
%}

\keyword{ package }

%\seealso{
%Optional links to other man pages, e.g.
%\code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}

%\examples{
%simple examples of the most important functions
%}
